chore: reorganise header-compte folder
